requirements.txt
discord.py>=2.3.0
postgrest-py>=0.5.0
python-dotenv>=1.0.0
